ocis-bin5: rename from ocis-bin
