+++ doc/release-notes/rl-2505.section.md
@@ -52,6 +52,8 @@
 
 - `i3status-rust`-package no longer enables `notmuch` by default. It can be enabled via `withNotmuch`.
 
+- `ocis-bin` has been renamed to `ocis_5-bin`. Future versions will have the major version suffix.
+
 - Default ICU version updated from 74 to 76
 
 - Apache Kafka was updated to `>= 4.0.0`. Please note that this is the first release which operates
